refactored, solved the isort problem
