+++ cdk/Pipfile
@@ -3,7 +3,7 @@ url = 'https://pypi.python.org/simple'
 verify_ssl = true
 
 [dev-packages]
-consoleme_ecs_service = {editable = true,path = "cdk"}
+consoleme_ecs_cdk = {editable = true,path = "consoleme_ecs_cdk"}
 
 [packages]
 boto3 = "*"
