+++ cdk/cdk.context.json
@@ -1,0 +1,7 @@
+{
+  "availability-zones:account=861504752911:region=eu-west-1": [
+    "eu-west-1a",
+    "eu-west-1b",
+    "eu-west-1c"
+  ]
+}
