+++ cdk/consoleme_ecs_cdk/pyproject.toml
