+++ cdk/consoleme_ecs_cdk/service/nested_stacks/shared_stack.py
