+++ cdk/consoleme_ecs_cdk/service/nested_stacks/vpc_stack.py
